PulmoVision-CNN-for-Pneumonia-Tuberculosis-Detection-main/requirements.txt
streamlit
gdown
pillow
torch
torchvision
torchaudio